修正了分页逻辑，添加了redis和ThreadLocal。增添了异常报错情况和修改了Jwt令牌处理的密码md5加密。

+++ yyzx_backend/src/main/java/com/cqupt/dto/CustomerPreferenceDTO.java
@@ -16,7 +16,7 @@ import lombok.EqualsAndHashCode;
 @ApiModel(value = "CustomerPreferenceDTO-顾客喜好查询条件", description = "")
 public class CustomerPreferenceDTO {
     @ApiModelProperty(value = "喜好编号")
-    private Integer customerPreferenceId;
+    private Long customerPreferenceId;
     @ApiModelProperty(value = "顾客姓名")
     private String customerName;
     @ApiModelProperty(value = "页码")
