+++ yyzx_backend/src/main/java/com/cqupt/dto/NurseRecordDTO.java
@@ -16,7 +16,7 @@ import lombok.EqualsAndHashCode;
 @ApiModel(value = "NurseRecordDTO-客户护理记录查询条件", description = "")
 public class NurseRecordDTO {
     @ApiModelProperty(value = "客户编号")
-    private Integer customerId;
+    private Long customerId;
     @ApiModelProperty(value = "页码")
     private Integer pageSize;
 }
