+++ yyzx_backend/src/main/java/com/cqupt/mapper/BackdownMapper.java
@@ -8,6 +8,6 @@ import org.apache.ibatis.annotations.Param;
 
 public interface BackdownMapper extends BaseMapper<Backdown> {
     Page<BackdownVo> selectBackdownVo(@Param("page") Page<BackdownVo> page,
-                                      @Param("userId") Integer userId
+                                      @Param("userId") Long userId
                                       ) throws Exception;
 }
