+++ yyzx_backend/src/main/java/com/cqupt/mapper/OutwardMapper.java
@@ -10,6 +10,6 @@ import org.apache.ibatis.annotations.Param;
 
 public interface OutwardMapper extends BaseMapper<Outward> {
     Page<OutwardVo> selectOutwardVo(@Param("page") Page<OutwardVo> page,
-                                    @Param("userId") Integer userId) throws Exception;
+                                    @Param("userId") Long userId) throws Exception;
 
 }
