+++ yyzx_backend/src/main/java/com/cqupt/pojo/CustomerPreference.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -25,10 +26,10 @@ public class CustomerPreference implements Serializable {
 
     @ApiModelProperty(value = "主键")
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty(value = "顾客ID")
-    private Integer customerId;
+    private Long customerId;
 
     @ApiModelProperty(value = "饮食喜好")
     private String preferences;
