+++ yyzx_backend/src/main/java/com/cqupt/service/OutwardService.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 public interface OutwardService extends IService<Outward> {
     ResultVo examineOutward(Outward  outward) throws Exception;
     ResultVo<Page<OutwardVo>> queryOutwardVo(OutwardDTO outwardDTO) throws Exception;
-    ResultVo delOutward(Integer id) throws Exception;
+    ResultVo delOutward(Long id) throws Exception;
     ResultVo updateOutward(Outward outward) throws Exception;
     ResultVo addOutward(Outward outward) throws Exception;
 }
